+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.12.0] - 2022-03-28
+
 ### Added
 - `genomepy.Annotation.lengths()` to retrieve the gene/transcript lengths.
 - `genomepy.Annotation.from_attributes()` can extract any sub-column that pesky attributes column
 
 ### Changed
-- updated Boyle-lab blacklists 
+- updated Boyle-lab blacklists
+- `genomepy.Annotation.genes()` default changed from bed (commonly containing transcript names) to gtf (gene names)
 
 ### Fixed
 - blacklists now work with GENCODE 
@@ -394,6 +397,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `-r` and `--match/--no-match` option to select sequences by regex.
 
 [Unreleased]: https://github.com/vanheeringen-lab/genomepy/compare/master...develop
+[0.12.0]: https://github.com/vanheeringen-lab/genomepy/compare/0.11.1...0.12.0
 [0.11.1]: https://github.com/vanheeringen-lab/genomepy/compare/0.11.0...0.11.1
 [0.11.0]: https://github.com/vanheeringen-lab/genomepy/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/vanheeringen-lab/genomepy/compare/0.9.3...0.10.0